Move numpy date shim into separate file
